# [js/webgpu] Destroy staging buffers aggressively during weights uploading